simplified db init and made suggested changes for tests
