+++ src/server/db/firebase.test.js
@@ -13,94 +13,41 @@ const mockdatabase = new firebasemock.MockFirebase();
 const mockfirestore = new firebasemock.MockFirestore();
 var mocksdk = new firebasemock.MockFirebaseSdk(
   // use null if your code does not use RTDB
-  path => {
-    return path ? mockdatabase.child(path) : mockdatabase;
-  },
+  () => mockdatabase,
   // use null if your code does not use AUTHENTICATION
   () => null,
   // use null if your code does not use FIRESTORE
-  () => {
-    return mockfirestore;
-  },
+  () => mockfirestore,
   // use null if your code does not use STORAGE
   () => null,
   // use null if your code does not use MESSAGING
   () => null
 );
 
-test('Firestore', async () => {
-  const mockConfig = {
+test('construction', () => {
+  const db = new Firebase({
+    mockFirebase: mocksdk,
+    dbname: 'a',
     engine: 'Firestore',
-    apiKey: 'apikey',
-    authDomain: 'authDomain',
-    databaseURL: 'databaseURL',
-    projectId: 'projectId',
-  };
-  mocksdk.initializeApp();
-  const mockFirebase = mocksdk.firestore();
-  mockFirebase.autoFlush();
-  const db = new Firebase({ config: mockConfig, mockFirebase });
-  await db.connect();
-
-  // Must return undefined when no game exists.
-  let state = await db.get('gameID');
-  expect(state).toEqual(null);
-
-  // Create game.
-  await db.set('gameID', { a: 1 });
-
-  // Cache hits.
-  {
-    // Must return created game.
-    state = await db.get('gameID');
-    expect(state).toMatchObject({ a: 1 });
-
-    // Must return true if game exists
-    const has = await db.has('gameID');
-    expect(has).toBe(true);
-  }
-
-  // Cache misses.
-  {
-    // Must return created game.
-    db.cache.reset();
-    state = await db.get('gameID');
-    expect(state).toMatchObject({ a: 1 });
-
-    // Must return true if game exists
-    db.cache.reset();
-    const has = await db.has('gameID');
-    expect(has).toBe(true);
-  }
-
-  // Cache size.
-  {
-    const db = new Firebase({ config: mockConfig, mockFirebase, cacheSize: 1 });
-    await db.connect();
-    await db.set('gameID', { a: 1 });
-    await db.set('another', { b: 1 });
-    state = await db.get('gameID');
-    // Check that it came from Mongo and not the cache.
-    expect(state._id).toBeDefined();
-  }
+  });
+  expect(db.dbname).toBe('a');
+  expect(db.config).toMatchObject({});
 });
 
-test('RTDB', async () => {
+test('Firestore', async () => {
   const mockConfig = {
-    engine: 'RTDB',
     apiKey: 'apikey',
     authDomain: 'authDomain',
     databaseURL: 'databaseURL',
     projectId: 'projectId',
   };
-  mocksdk.initializeApp();
-  const mockFirebase = mocksdk.database();
-  mockFirebase.ref().autoFlush();
   const db = new Firebase({
     config: mockConfig,
-    mockFirebase: mockFirebase.ref(),
+    mockFirebase: mocksdk,
+    engine: 'Firestore',
   });
   await db.connect();
+  db.db.autoFlush();
 
   // Must return undefined when no game exists.
   let state = await db.get('gameID');
@@ -137,8 +84,75 @@ test('RTDB', async () => {
   {
     const db = new Firebase({
       config: mockConfig,
-      mockFirebase: mockFirebase.ref(),
+      mockFirebase: mocksdk,
       cacheSize: 1,
+      engine: 'Firestore',
+    });
+    await db.connect();
+    await db.set('gameID', { a: 1 });
+    await db.set('another', { b: 1 });
+    state = await db.get('gameID');
+    // Check that it came from Mongo and not the cache.
+    expect(state._id).toBeDefined();
+  }
+});
+
+test('RTDB', async () => {
+  const mockConfig = {
+    apiKey: 'apikey',
+    authDomain: 'authDomain',
+    databaseURL: 'databaseURL',
+    projectId: 'projectId',
+  };
+  // mocksdk.initializeApp();
+  // const mockFirebase = mocksdk.database();
+  // mockFirebase.ref().autoFlush();
+  const db = new Firebase({
+    config: mockConfig,
+    mockFirebase: mocksdk,
+    engine: 'RTDB',
+  });
+  await db.connect();
+  db.db.autoFlush();
+
+  // Must return undefined when no game exists.
+  let state = await db.get('gameID');
+  expect(state).toEqual(null);
+
+  // Create game.
+  await db.set('gameID', { a: 1 });
+
+  // Cache hits.
+  {
+    // Must return created game.
+    state = await db.get('gameID');
+    expect(state).toMatchObject({ a: 1 });
+
+    // Must return true if game exists
+    const has = await db.has('gameID');
+    expect(has).toBe(true);
+  }
+
+  // Cache misses.
+  {
+    // Must return created game.
+    db.cache.reset();
+    state = await db.get('gameID');
+    expect(state).toMatchObject({ a: 1 });
+
+    // Must return true if game exists
+    db.cache.reset();
+    const has = await db.has('gameID');
+    expect(has).toBe(true);
+  }
+
+  // Cache size.
+  {
+    const db = new Firebase({
+      config: mockConfig,
+      mockFirebase: mocksdk,
+      cacheSize: 1,
+      engine: 'RTDB',
     });
     await db.connect();
     await db.set('gameID', { a: 1 });
@@ -151,18 +165,21 @@ test('RTDB', async () => {
 
 test('Firestore - race conditions', async () => {
   const mockConfig = {
-    engine: 'Firestore',
     apiKey: 'apikey',
     authDomain: 'authDomain',
     databaseURL: 'databaseURL',
     projectId: 'projectId',
   };
-  mocksdk.initializeApp();
-  const mockFirebase = mocksdk.firestore();
-  mockFirebase.autoFlush();
-  const db = new Firebase({ config: mockConfig, mockFirebase });
+  // mocksdk.initializeApp();
+  // const mockFirebase = mocksdk.firestore();
+  // mockFirebase.autoFlush();
+  const db = new Firebase({
+    config: mockConfig,
+    mockFirebase: mocksdk,
+    engine: 'Firestore',
+  });
   await db.connect();
-
+  db.db.autoFlush();
   // Out of order set()'s.
   await db.set('gameID', { _stateID: 1 });
   await db.set('gameID', { _stateID: 0 });
@@ -183,21 +200,21 @@ test('Firestore - race conditions', async () => {
 
 test('RTDB - race conditions', async () => {
   const mockConfig = {
-    engine: 'RTDB',
     apiKey: 'apikey',
     authDomain: 'authDomain',
     databaseURL: 'databaseURL',
     projectId: 'projectId',
   };
-  mocksdk.initializeApp();
-  const mockFirebase = mocksdk.database();
-  mockFirebase.ref().autoFlush();
+  // mocksdk.initializeApp();
+  // const mockFirebase = mocksdk.database();
+  // mockFirebase.ref().autoFlush();
   const db = new Firebase({
     config: mockConfig,
-    mockFirebase: mockFirebase.ref(),
+    mockFirebase: mocksdk,
+    engine: 'RTDB',
   });
   await db.connect();
-
+  db.db.autoFlush();
   // Out of order set()'s.
   await db.set('gameID', { _stateID: 1 });
   await db.set('gameID', { _stateID: 0 });
