+++ src/server/index.js
@@ -10,7 +10,7 @@ const Koa = require('koa');
 const IO = require('koa-socket');
 const Redux = require('redux');
 
-import { InMemory, Mongo, Firebase } from './db';
+import { DBFromEnv } from './db';
 import { CreateGameReducer } from '../core/reducer';
 import { createApiServer, isActionFromAuthenticPlayer } from './api';
 
@@ -28,27 +28,7 @@ export function Server({ games, db, _clientInfo, _roomInfo }) {
   app.context.io = io;
   io.attach(app);
 
-  if (db === undefined) {
-    if (process.env.MONGO_URI) {
-      db = new Mongo({ url: process.env.MONGO_URI });
-    } else if (
-      process.env.FIREBASE_APIKEY &&
-      process.env.FIREBASE_AUTHDOMAIN &&
-      process.env.FIREBASE_DATABASEURL &&
-      process.env.FIREBASE_PROJECTID
-    ) {
-      const config = {
-        apiKey: process.env.FIREBASE_APIKEY,
-        authDomain: process.env.FIREBASE_AUTHDOMAIN,
-        databaseURL: process.env.FIREBASE_DATABASEURL,
-        projectId: process.env.FIREBASE_PROJECTID,
-        engine: process.env.FIREBASE_ENGINE, // RTDB or Firestore
-      };
-      db = new Firebase({ config });
-    } else {
-      db = new InMemory();
-    }
-  }
+  if (db === undefined) db = DBFromEnv();
 
   const api = createApiServer({ db, games });
